+++ apps/publish/src/release.ts
@@ -1,9 +1,11 @@
 import "dotenv/config"
 import pg from "pg"
 import cleanEntities from "./make/cleanEntities.js"
+import deployWww from "./make/deployWww.js"
 import getBuild from "./make/getBuild.js"
 import invalidateAPICache from "./make/invalidateAPICache.js"
 import updateParameters from "./make/updateParameters.js"
+import updateWwwBuildEnv from "./make/updateWwwBuildEnv.js"
 ;(async () => {
     const pgClient = new pg.Client({
         database: "phylopic-entities",
@@ -14,11 +16,13 @@ import updateParameters from "./make/updateParameters.js"
         console.info("Releasing build", build, "...")
         await updateParameters(build /*, source.root*/)
         console.info("Build", build, "released.")
+        updateWwwBuildEnv(build)
         await Promise.all([
             (async () => {
                 console.info("Invalidating API cache...")
                 await invalidateAPICache(build)
                 console.info("Invalidated API cache.")
+                await deployWww(build)
             })(),
             (async () => {
                 console.info("Cleaning up entities database...")