+++ apps/publish/CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
+### Security
+
+## [1.15.0] - 2026-09-13
+
+### Added
+
 - `yarn release` sets Vercel `NEXT_PUBLIC_BUILD` on all environments and deploys `www`.
 - `yarn release` writes `NEXT_PUBLIC_BUILD` to `apps/www/.env.local`.